A test of authenticated? with a nonexistent digest. red
The second case, involving a scenario with two different browsers, is
harder to simulate with an integration test, but it’s easy to check in
the User model test directly. All we need is to start with a user that
has no remember digest (which is true for the @user variable defined in
the setup method) and then call authenticated?, as shown in this commit.
(Note that we’ve just left the remember token blank; it doesn’t matter
what its value is, because the error occurs before it ever gets used.)

Because BCrypt::Password.new(nil) raises an error, the test suite
should now be red:

$ rails test

ERROR["test_authenticated?_should_return_false_for_a_user_with_nil_diges
                                      t", UserTest, 14.529656675003935]
 test_authenticated?_should_return_false_for_a_user_with_nil_digest#User
                                                          Test (14.53s)
BCrypt::Errors::InvalidHash:         BCrypt::Errors::InvalidHash:
                                                            invalid hash
            app/models/user.rb:32:in `new'
            app/models/user.rb:32:in `authenticated?'
            test/models/user_test.rb:77:in `block in <class:UserTest>'