Updating authenticated? to handle a nonexistent digest. green
To fix the error and get to green, all we need to do is return false if
the remember digest is nil

This uses the return keyword to return immediately if the remember
digest is nil, which is a common way to emphasize that the rest of the
method gets ignored in that case. The equivalent code

if remember_digest.nil?
  false
else
  BCrypt::Password.new(remember_digest).is_password?(remember_token)
end
would also work fine, but I prefer the explicitness of the version in
commit: Updating authenticated? to handle a nonexistent digest
(which also happens to be slightly shorter).

With the code in commit: Updating authenticated? to handle a nonexistent
digest, our full test suite should be green, and both subtleties should
now be addressed:

$ rails test

Exercises -

1. Comment out the fix in commit: Only logging out if logged in and then
verify that the first subtle bug is present by opening two logged-in
tabs, logging out in one, and then clicking “Log out” link in the other.

Yes, here is what happens when this is done.

Started DELETE "/logout" for 127.0.0.1 at 2017-12-09 23:31:54 -0500
Processing by SessionsController#destroy as HTML
  Parameters: {"authenticity_token"=>"Mu87rl11vlap6ZOsgllW3VfRIIgsefXhms
                4WqaxnxGaP9hbItW3nJukg0JKfmZAAC1khVx2EfEOII+DroU8orQ=="}
Completed 500 Internal Server Error in 1ms (ActiveRecord: 0.0ms)

NoMethodError (undefined method `forget' for nil:NilClass):

app/helpers/sessions_helper.rb:34:in `forget'
app/helpers/sessions_helper.rb:41:in `log_out'
app/controllers/sessions_controller.rb:18:in `destroy'

2. Comment out the fix in commit: Updating authenticated? to handle a
nonexistent digest. and verify that the second
subtle bug is present by logging out in one browser and closing and
opening the second browser.

Its a little difficult to demo this bug but I feel comfortable to moving
on

3. Uncomment the fixes and confirm that the test suite goes from red to
green.

Done.