Adding a “remember me” checkbox to the login form.
With the code in previous commits, our application has a complete,
professional-grade authentication system. As a final step, we’ll see
how to make staying logged in optional using a “remember me” checkbox.

To write the implementation, we start by adding a checkbox to the login
form from commit: Code for the login form. As with labels,
text fields, password fields, and submit buttons, checkboxes can be
created with a Rails helper method. In order to get the styling right,
though, we have to nest the checkbox inside the label, as follows: