Handling the submission of the “remember me” checkbox.
params[:session][:remember_me]

is ’1’ if the box is checked and ’0’ if it isn’t.

By testing the relevant value of the params hash, we can now remember
or forget the user based on the value of the submission:15

if params[:session][:remember_me] == '1'
  remember(user)
else
  forget(user)
end

This sort of if-then branching structure can be converted to one line
using the ternary operator as follows:

params[:session][:remember_me] == '1' ? remember(user) : forget(user)

Using this to replace remember user in the Sessions controller’s create
method, commit: Logging in and remembering a use, leads to the amazingly
compact code

(Now you’re in a position to understand the code in commit:
Adding a digest method for use in fixtures, which
uses the ternary operator to define the bcrypt cost variable.)

With the implementation in this commit, our login system is complete,
as you can verify by checking or unchecking the box in your browser.

Exercises -

1. By inspecting your browser’s cookies directly, verify that the
“remember me” checkbox is having its intended effect.

Yes, the cookie values are present only remember is checked

2. At the console, invent examples showing both possible behaviors of
the ternary operator.

Not sure how demo this.