A test of the “remember me” checkbox. green
When I originally implemented the checkbox handling in commit: Handling
the submission of the “remember me” checkbox., instead of the correct

params[:session][:remember_me] == '1' ? remember(user) : forget(user)
I actually used

params[:session][:remember_me] ? remember(user) : forget(user)
In this context, params[:session][:remember_me] is either ’0’ or ’1’,
both of which are true in a boolean context, so the resulting expression
is always true, and the application acts as if the checkbox is always
checked. This is exactly the kind of error a test can catch.

Because remembering users requires that they be logged in, our first
step is to define a helper to log users in inside tests. In commit: A
test for user logging in with valid information., we logged a user in
using the post method and a valid session hash, but it’s cumbersome to
do this every time. To avoid needless repetition, we’ll write a helper
method called log_in_as to log in for us.

Our method for logging a user in depends on the type of test. Inside
controller tests, we can manipulate the session method directly,
assigning user.id to the :user_id key (as first seen in commit: The
log_in function):

def log_in_as(user)
  session[:user_id] = user.id
end

We call the method log_in_as to avoid any confusion with the
application code’s log_in method as defined in commit: The
log_in function. Its location is in the ActiveSupport::TestCase class
inside the test_helper file, the same location as the is_logged_in?
helper from commit:A boolean method for login status inside tests:

class ActiveSupport::TestCase
  fixtures :all

  # Returns true if a test user is logged in.
  def is_logged_in?
    !session[:user_id].nil?
  end

  # Log in as a particular user.
  def log_in_as(user)
    session[:user_id] = user.id
  end
end
We won’t actually need this version of the method in this chapter, but
we’ll put it to use in Chapter 10.

Inside integration tests, we can’t manipulate session directly, but we
can post to the sessions path as in commit: A test for user logging in
with valid information, which leads to the log_in_as method shown here:

class ActionDispatch::IntegrationTest

  # Log in as a particular user.
  def log_in_as(user, password: 'password', remember_me: '1')
    post login_path, params: { session: { email: user.email,
                                          password: password,
                                          remember_me: remember_me } }
  end
end

Because it’s located inside the ActionDispatch::IntegrationTest class,
this is the version of log_in_as that will be called inside integration
tests. We use the same method name in both cases because it lets us do
things like use code from a controller test in an integration without
making any changes to the login method.

Putting these two methods together yields the parallel log_in_as helpers
shown in commit:Adding a log_in_as helper.