A test of the “remember me” checkbox. green
After logging in, we can check if the user has been remembered by
looking for the remember_token key in the cookies. Ideally, we would
check that the cookie’s value is equal to the user’s remember token,
but as currently designed there’s no way for the test to get access to
it: the user variable in the controller has a remember token attribute,
but (because remember_token is virtual) the @user variable in the test
doesn’t. Fixing this minor blemish is left as an exercise but for now
we can just test to see if the relevant cookie is nil or not.

There’s one more subtlety, which is that for some reason inside tests
the cookies method doesn’t work with symbols as keys, so that

cookies[:remember_token]

is always nil. Luckily, cookies does work with string keys, so that

cookies['remember_token']

has the value we need.

(Recall that users(:michael) references the fixture user from previous)