Exercises - access the virtual remember_token attr for integration test
As mentioned above, the application currently doesn’t have any way to
access the virtual remember_token attribute in the integration test in
commit: A test of the “remember me” checkbox.. It is possible, though,
using a special test methodcalled assigns. Inside a test, you can access
instance variables defined in the controller by using assigns with the
corresponding symbol. For example, if the create action defines an @user
variable, we can access it in the test using assigns(:user). Right now,
the Sessions controller create action defines a normal (non-instance)
variable called user, but if we change it to an instance variable we
can test that cookies correctly contains the user’s remember token. By
filling in the missing elements in commit: A template for using an
instance variable in the create action.  and in commit: A template for
an improved “remember me” test. (indicated with question marks ?
and FILL_IN), complete this improved test of the “remember me” checkbox.