Raising an exception in an untested branch.
we verified by hand that the persistent session implemented in the
preceding sections is working, but in fact the relevant branch in the
current_user method is currently completely untested. My favorite way
to handle this kind of situation is to raise an exception in the
suspected untested block of code: if the code isn’t covered, the tests
will still pass; if it is covered, the resulting error will identify
the relevant test. The result in the present case appears in this
commit.

$ rails test