The user edit view.
The corresponding user edit view (which you will have to create by hand)
in this commit. Note how closely this resembles the new user
view from commit: A form to sign up new users; the large overlap
suggests factoring the repeated code into a partial, which is left as
up coming exercise.

Here we have reused the shared error_messages partial introduced in
commit: A partial for displaying form submission error messages.
By the way, the use of target="_blank" in the Gravatar link is a neat
trick to get the browser to open the page in a new
window or tab, which is sometimes convenient behavior when linking to
third-party sites. (There’s a minor security issue associated with
target="_blank"; dealing with this detail is left as an exercise.

With the @user instance variable from commit: The user edit action,
the edit page should render properly. The “Name” and “Email” fields
in within this commit also shows how Rails automatically pre-fills the
Name and Email fields using the attributes of the existing @user
variable.

Since web browsers can’t natively send PATCH requests (as required by
the REST conventions from table in commit: The user edit action), Rails
fakes it with a POST request and a hidden input field.

There’s another subtlety to address here: the code form_for(@user)
in this commit is exactly the same as the code in commit: A form to
sign up new users—so how does Rails know to use a POST request for new
users and a PATCH for editing users? The answer is that it is possible
to tell whether a user is new or already exists in the database via
Active Record’s new_record? boolean method:

$ rails console
>> User.new.new_record?
=> true
>> User.first.new_record?
=> false

When constructing a form using form_for(@user), Rails uses POST if
@user.new_record? is true and PATCH if it is false.