Exercises - minor security issue associated URL
As noted above, there’s a minor security issue associated with using
target="_blank" to open URLs, which is that the target site gains
control of what’s known as the “window object” associated with the HTML
document. The result is that the target site could potentially introduce
malicious content, such as a phishing page. This is extremely unlikely
to happen when linking to a reputable site like Gravatar, but it turns
out that we can eliminate the risk entirely by setting the rel attribute
(“relationship”) to "noopener" in the origin link. Add this attribute
to the Gravatar edit link in Listing 10.2.