The edit view with partial.
Remove the duplicated form code by refactoring the new.html.erb and
edit.html.erb views to use the partial in commit: A partial for the new
and edit form, as shown in commit: The signup view with partiat and this
commit.

Note the use of the provide method, which we used before in Section:
Layouts and embedded Ruby (Refactor) to eliminate duplication in the
layout.

(If you solved the exercise corresponding to commit: Posting to the
/signup URL, this exercise won’t work exactly as written, and you’ll
have to apply your technical sophistication to resolve the discrepancy.
My suggestion is to use the variable-passing technique shown in
commit: A partial for the new and edit form to pass the needed URL
from commit: The signup view with partial and in commit:The edit view
with partial to the form in commit: A partial for the new and edit form)