The initial user update action
In this commit we’ll handle unsuccessful edits,
following similar ideas to unsuccessful signups. We start by creating an
update action, which uses update_attributes to update the user based on
the submitted params hash, as shown in Listing 10.8. With invalid
information, the update attempt returns false, so the else branch
renders the edit page. We’ve seen thispattern before; the structure
closely parallels the first version of the create action in the commit:
A create action that can handle signup failure.

Note the use of user_params in the call to update_attributes, which uses
strong parameters to prevent mass assignment vulnerability (as described
in Section 7.3.2).

Because of the existing User model validations and the error-messages
partial in commit: The user edit view, submission of invalid information
results in helpful error messages.

Exercises -

Confirm by submitting various invalid combinations of username, email,
and password that the edit form won’t accept invalid submissions.

Yes, the various invalid combinations for username, email,
and password does appear. Below are the errors message that I see.

The form contains 5 errors.
  Name can't be blank
  Email can't be blank
  Email is invalid
  Password can't be blank
  Password is too short (minimum is 6 characters)