Exercises -
Add a line in commit: A test for an unsuccessful edit. to test for the
correct number of error messages. Hint: Use an assert_select (review
README.md: Some useful assert_select options) that tests for a div with
class alert containing the text “The form contains 4 errors.”

The solution is in this commit.