A test of a successful edit. red
As you get more comfortable with testing, you might find that it’s
useful to write integration tests before writing the application code
instead of after. In this context, such tests are sometimes known as
acceptance tests, since they determine when a particular feature should
be accepted as complete. To see how this works, we’ll complete the user
edit feature using test-driven development.

We’ll test for the correct behavior of updating users by writing a test
similar to the one shown in commit: A test for an unsuccessful edit,
only this time we’ll submit valid information. Then we’ll check for a
nonempty flash message and a successful redirect to the profile page,
while also verifying that the user’s information correctly changed in
the database. The result appears in commit: A test of a successful edit.
Note that the password and confirmation in commit: A test of a
successful edit, are blank, which is convenient for users who don’t want
to update their passwords every time they update their names or email
addresses. Note also the use of @user.reload (first seen in
Section 6.1.5 Updating user objects) to reload the user’s values from
the database and confirm that they were successfully updated. (This is
the kind of detail you could easily forget initially, which is why
acceptance testing (and TDD generally) require a certain level of
experience to be effective.)

After running:
 $rails test

 FAIL["test_successful_edit", UsersEditTest, 20.957933499012142]
 test_successful_edit#UsersEditTest (20.96s)
        Expected true to be nil or false
        test/integration/users_edit_test.rb:30:in `block in
                                                  <class:UsersEditTest>'