Allowing empty passwords on update. green
As indicated in the caption to in commit: The user update action,
the test suite is still red, which is the result of the password length
validation (commit: The complete implementation for secure passwords.)
failing due to the empty password and confirmation in commit: A test
of a successful edit. To get the tests to green, we need to make an
exception to the password validation if the password is empty. We can
do this by passing the allow_nil: true option to validates, as seen in
this commit.

In case you’re worried that this commit might allow new users to sign
up with empty passwords, recall from Section 6.3.3 Minimum password
standards that has_secure_password includes a separate presence
validation that specifically catches nil passwords. (Because nil
passwords now bypass the main presence validation but are still caught
by has_secure_password, this also fixes the duplicate error message
mentioned in Section 7.3.3 Signup error messages.)

With the code in this commit, the user edit page should be working,
as you can double-check by re-running the test suite, which should now
be green:

$ rails test

Exercises -

Double-check that you can now make edits by making a few changes on the
development version of the application.

Yes, everything does work.

What happens when you change the email address to one without an
associated Gravatar?

We get a standard gravatar that does have gravatar that is known.