Adding a logged_in_user before filter. red
In the context of web applications, authentication allows us to identify
users of our site, while authorization lets us control what they can do.
One nice effect of building the authentication machinery in Chapter 8 is
that we are now in a position to implement authorization as well.

Although the edit and update actions from commit: The user edit action
are functionally complete, they suffer from a ridiculous security flaw:
they allow anyone (even non-logged-in users) to access either action and
update the information for any user. In this section, we’ll implement a
security model that requires users to be logged in and prevents them
from updating any information other than their own.

In this commit, we’ll handle the case of non-logged-in users who try to
access a protected page to which they might normally have access.
Because this could easily happen in the normal course of using the
application, such users will be forwarded to the login page with a
helpful message. On the other hand, users who try to access a page for
which they would never be authorized (such as a logged-in user trying to
access a different user’s edit page) will be redirected to the root URL
(commit: Adding a second user to the fixture file.).

To implement the forwarding behavior shown in Figure 10.6, we’ll use a
before filter in the Users controller. Before filters use the
before_action command to arrange for a particular method to be called
before the given actions. To require users to be logged in, we define a
logged_in_user method and invoke it using before_action :logged_in_user,
as shown in this commit.

By default, before filters apply to every action in a controller, so here we restrict the filter to act only on the :edit and :update actions by passing the appropriate only: options hash.

$ rails test

The reason is that the edit and update actions now require a logged-in user, but no user is logged in inside the corresponding tests.

# Running:

 FAIL["test_successful_edit", UsersEditTest, 14.186544398999104]
 test_successful_edit#UsersEditTest (14.19s)
        expecting <"users/edit"> but rendering with <[]>
        test/integration/users_edit_test.rb:23:in `block in <class:UsersEditTest>'

==F

Failure:
UsersEditTest#test_successful_edit [C:/Users/camcgruder/git/sample_app_5/test/integration/users_edit_test.rb:23]:
expecting <"users/edit"> but rendering with <[]>

bin/rails test test/integration/users_edit_test.rb:21

 FAIL["test_unsuccessful_edit", UsersEditTest, 14.197435334001057]
 test_unsuccessful_edit#UsersEditTest (14.20s)
        expecting <"users/edit"> but rendering with <[]>
        test/integration/users_edit_test.rb:11:in `block in <class:UsersEditTest>'