Commenting out the before filter to test our security model. green
Even though our test suite is now passing, we’re not finished with the before filter, because the suite is still green even if we remove our security model, as you can verify by commenting it out (commit: Commenting out the before filter to test our security model). This is a Bad Thing—of all the regressions we’d like our test suite to catch, a massive security hole is probably #1, so the code in commit: Commenting out the before filter to test our security model should definitely be red. Let’s write tests to arrange that.