Adding a second user to the fixture file
Of course, requiring users to log in isn’t quite enough; users should
only be allowed to edit their own information. It’s easy to have a test
suite that misses an essential security flaw, so we’ll proceed using
test-driven development to be sure our code implements the security
model correctly. To do this, we’ll add tests to the Users controller
test to complement the ones shown in Listing 10.20.

In order to make sure users can’t edit other users’ information, we need
to be able to log in as a second user. This means adding a second user
to our users fixture file, in this commit.