Tests for trying to edit as the wrong user. red
By using the log_in_as method defined in commit: Adding a log_in_as
helper, we can test the edit and update actions as in commit. Note that
we expect to redirect users to the root path instead of the login path
because a user trying to edit a different user would already be logged
in.