A before filter to protect the edit/update pages. green
To redirect users trying to edit another user’s profile, we’ll add a
second method called correct_user, together with a before filter to call
it (in this commit). Note that the correct_user before filter defines
the @user variable, so in this commit also shows that we can eliminate
the @user assignments in the edit and update actions.