The final correct_user before filter. green
Replacing the direct comparison with the Boolean method gives the code
in commit

Exercises -

Why is it important to protect both the edit and update actions?

It's important to make sure the correct person can edit and update the
users data.

Which action could you more easily test in a browser?

show, new, and edit because they have views associated with them.