A test for friendly forwarding. red
Our site authorization is complete as written, but there is one minor
blemish: when users try to access a protected page, they are currently
redirected to their profile pages regardless of where they were trying
to go. In other words, if a non-logged-in user tries to visit the edit
page, after logging in the user will be redirected to /users/1 instead
of /users/1/edit. It would be much friendlier to redirect them to their
intended destination instead.

The application code will turn out to be relatively complicated, but we
can write a ridiculously simple test for friendly forwarding just by
reversing the order of logging in and visiting the edit page in commit:
Logging in a test user. As seen in this commit, the resulting test tries
to visit the edit page, then logs in, and then checks that the user is
redirected to the edit page instead of the default profile page. (In
this commit also removes the test for rendering the edit template since
that’s no longer the expected behavior.)