Code to implement friendly forwarding. red
Now that we have a failing test, we’re ready to implement friendly
forwarding. In order to forward users to their intended destination, we
need to store the location of the requested page somewhere, and then
redirect to that location instead of to the default. We accomplish this
with a pair of methods, store_location and redirect_back_or, both
defined in the Sessions helper

 FAIL["test_successful_edit_with_friendly_forwarding", UsersEditTest,
20.1140191279992]
 test_successful_edit_with_friendly_forwarding#UsersEditTest (20.11s)
        Expected response to be a redirect to
<http://www.example.com/users/762146111/edit> but was a redirect to
<http://www.example.com/users/762146111>.
        Expected "http://www.example.com/users/762146111/edit" to be ===
"http://www.example.com/users/762146111".
        test/integration/users_edit_test.rb:42:in `block in
<class:UsersEditTest>'

==F