Adding store_location to the logged-in user before filter. red
 Here the storage mechanism for the forwarding URL is the same session
facility we used in commit: The log_in function to log the user in.
Listing 10.30 also uses the request object (via request.original_url) to
get the URL of the requested page.

The store_location method in commit: Code to implement friendly
forwarding puts the requested URL in the session variable under the key
:forwarding_url, but only for a GET request. This prevents storing the
forwarding URL if a user, say, submits a form when not logged in (which
is an edge case but could happen if, e.g., a user deleted the session
cookies by hand before submitting the form). In such a case, the
resulting redirect would issue a GET request to a URL expecting POST,
PATCH, or DELETE, thereby causing an error. Including if request.get?
prevents this from happening.6

To make use of store_location, we need to add it to the logged_in_user
before filter, as shown in this commit.