The Sessions create action with friendly forwarding. green
To implement the forwarding itself, we use the redirect_back_or method
to redirect to the requested URL if it exists, or some default URL
otherwise, which we add to the Sessions controller create action to
redirect after successful login (this commit). The redirect_back_or
method uses the or operator || through

session[:forwarding_url] || default

This evaluates to session[:forwarding_url] unless it’s nil, in which
case it evaluates to the given default URL. Note that commit: Code to
implement friendly forwarding is careful to remove the forwarding URL
(via session.delete(:forwarding_url)); otherwise, subsequent login
attempts would forward to the protected page until the user closed their
browser. (Testing for this is left as an exercise. Also note that the
session deletion occurs even though the line with the redirect appears
first; redirects don’t happen until an explicit return or the end of the
method, so any code appearing after the redirect is still executed.