The Sessions create action with friendly forwarding. green
With that, the friendly forwarding integration test in Listing 10.29
should pass, and the basic user authentication and page protection
implementation is complete. As usual, it’s a good idea to verify that
the test suite is green before proceeding:

$ rails test

Exercises -

Write a test to make sure that friendly forwarding only forwards to the
given URL the first time. On subsequent login attempts, the forwarding
URL should revert to the default (i.e., the profile page). Hint: Add to
the test in Listing 10.29 by checking for the right value of
session[:forwarding_url].

Used the link below figure out solution:

https://stackoverflow.com/questions/30768657/how-do-i-check-to-see-if-a-user-will-be-redirected-to-his-her-profile-after-logi

Put a debugger in the Sessions controller’s new action, then log out and
try to visit /users/1/edit. Confirm in the debugger that the value of
session[:forwarding_url] is correct. What is the value of request.get?
for the new action? (Sometimes the terminal can freeze up or act
strangely when you’re using the debugger; use your technical
sophistication (Box 1.1) to resolve any issues.)