Testing the index action redirect. red
In this section, we’ll add the penultimate user action, the index
action,
which is designed to display all the users instead of just one. Along
the way, we’ll learn how to seed the database with sample users and how
to paginate the user output so that the index page can scale up to
display a potentially large number of users. A mockup of the
result—users, pagination links, and a “Users” navigation link—appears in
[Figure
10.8.7](https://www.railstutorial.org/book/updating_and_deleting_users#fig-user_index_mockup)

In Section 10.4, we’ll add an administrative interface to the users
index so that users can also be destroyed.

To get started with the users index, we’ll first implement a security
model. Although we’ll keep individual user show pages visible to all
site visitors, the user index will be restricted to logged-in users so
that there’s a limit to how much unregistered users can see by default.

To protect the index page from unauthorized access, we’ll first add a
short test to verify that the index action is redirected properly
(in this commit).

ERROR["test_should_redirect_index_when_not_logged_in",
UsersControllerTest, 18.380293992002407]
 test_should_redirect_index_when_not_logged_in#UsersControllerTest
(18.38s)
AbstractController::ActionNotFound:
AbstractController::ActionNotFound: The action 'index' could not be
found for UsersController
            test/controllers/users_controller_test.rb:11:in `block in
<class:UsersControllerTest>'