The user index action.
To display the users themselves, we need to make a variable containing
all the site’s users and then render each one by iterating through them
in the index view. As you may recall from the corresponding action in
the toy app, we can use User.all to pull all the users out of the
database, assigning them to an @users instance variable for use in the
view, as seen in Listing 10.36. (If displaying all the users at once
seems like a bad idea, you’re right, and we’ll remove this blemish)