Adding an options hash in the gravatar_for helper.
The code in previous commit uses the result of this commit from commit:
Adding an options hash in the gravatar_for helper, which allows us to
pass an option to the Gravatar helper specifying a size other than the
default. If you didn’t do that exercise, update your Users helper file
with the contents of this commit before proceeding. (You are also
welcome to use the Ruby 2.0–style version from commit: Using keyword
arguments in the gravatar_for helper instead.)