Exercises - filled in all the links in the site layout
We’ve now filled in all the links in the site layout. Write an integration test for all the layout links, including the proper behavior for logged-in and non-logged-in users. Hint: Use the
log_in_as helper and add to the steps shown in commit: A test
for the links on the layout

Got some help from:

https://www.railstutorial.org/book/updating_and_deleting_users#fig-user_index_mockup