Adding the Faker gem to the Gemfile
In this section, we’ll give our lonely sample user some company. Of
course, to create enough users to make a decent users index, we could
use our web browser to visit the signup page and make the new users one
by one, but a far better solution is to use Ruby to make the users for
us.

First, we’ll add the Faker gem to the Gemfile, which will allow us to
make sample users with semi-realistic names and email addresses in this
commit (Ordinarily, you’d probably want to restrict the faker gem to a
development environment, but in the case of the sample app we’ll be
using it on our production site as well.)

Then install as usual:

$ bundle install