The users index with pagination.
To get pagination working, we need to add some code to the index view
telling Rails to paginate the users, and we need to replace User.all in
the index action with an object that knows about pagination. We’ll start
by adding the special will_paginate method in the view (in this commit);
we’ll see in a moment why the code appears both above and below the user
list.

The will_paginate method is a little magical; inside a users view, it
automatically looks for an @users object, and then displays pagination
links to access other pages. The view in commit: The users index with
pagination doesn’t work yet, though, because currently @users contains
the results of User.all (in commit: The user index action), whereas
will_paginate requires that we paginate the results explicitly using the
paginate method:

$ rails console

```ruby
>> User.paginate(page: 1)
  User Load (1.5ms)  SELECT "users".* FROM "users" LIMIT 30 OFFSET 0
   (1.7ms)  SELECT COUNT(*) FROM "users"
=> #<ActiveRecord::Relation [#<User id: 1,...
```

Note that paginate takes a hash argument with key :page and value equal
to the page requested. User.paginate pulls the users out of the database
one chunk at a time (30 by default), based on the :page parameter. So,
for example, page 1 is users 1–30, page 2 is users 31–60, etc. If page
is nil, paginate simply returns the first page.