Paginating the users in the index action.
Using the paginate method, we can paginate the users in the sample
application by using paginate in place of all in the index action. Here
the page parameter comes from params[:page], which is generated
automatically by will_paginate.

The users index page should now be working, appearing as in [Figure
10.11](https://www.railstutorial.org/book/updating_and_deleting_users#fig-user_index_pagination).
(On some systems, you may have to restart the Rails server at this
point.) Because we included will_paginate both above and below the user
list, the pagination links appear in both places.

If you now click on either the 2 link or Next link, you’ll get the
second page of results, as shown in [Figure
10.12](https://www.railstutorial.org/book/updating_and_deleting_users#fig-user_index_page_two_rails_3).

Exercises -

Confirm at the console that setting the page to nil pulls out the first
page of users.

First check normal ActiveRecord call to requset all users

```ruby
$ rails console
>> User.all
User.all
  User Load (26.6ms)  SELECT  "users".* FROM "users" LIMIT ?  [["LIMIT",
11]]
=> #<ActiveRecord::Relation ...
>> User.paginate(page: 1)
User.paginate(page: 1)
  User Load (22.6ms)  SELECT  "users".* FROM "users" LIMIT ? OFFSET ?
[["LIMIT", 11], ["OFFSET", 0]]
   (0.5ms)  SELECT COUNT(*) FROM "users"
=> #<ActiveRecord::Relation ...
>>  User.paginate(page: 2)
User.paginate(page: 2)
  User Load (0.5ms)  SELECT  "users".* FROM "users" LIMIT ? OFFSET ?
[["LIMIT", 11], ["OFFSET", 11]]
   (0.5ms)  SELECT COUNT(*) FROM "users"
=> #<ActiveRecord::Relation ...
```

What is the Ruby class of the pagination object? How does it compare to
the class of User.all?

We can look at the calls above, that we are receiving an
ActiveRecord::Relation object, which is a the same object that is
returned in User.all.