Adding 30 extra users to the fixture.
Now that our users index page is working, we’ll write a lightweight test
for it, including a minimal test for the pagination from Section 10.3.3:
Pagination. The idea is to log in, visit the index path, verify the
first page of users is present, and then confirm that pagination is
present on the page. For these last two steps to work, we need to have
enough users in the test database to invoke pagination, i.e., more than
30.

We created a second user in the fixtures in commit: Adding a second user
to the fixture file., but 30 or so more users is a lot to create by
hand. Luckily, as we’ve seen with the user fixture’s password_digest
attribute, fixture files support embedded Ruby, which means we can
create 30 additional users as shown in this commit. (In this commit also
creates a couple of other named users for future reference.)