A test of the users index, including pagination. green
With the fixtures defined in commit: Adding 30 extra users to the
fixture, we’re ready to write a test of the users index. First we
generate the relevant test:

```ruby
$ rails generate integration_test users_index
      invoke  test_unit
      create    test/integration/users_index_test.rb
```

The test itself involves checking for a div with the required pagination
class and verifying that the first page of users is present. The result
appears in this commit.

The result should be a green test suite:

$ rails test

Exercises -

By commenting out the pagination links in commit: The users index with
pagination, confirm that the test in this commit goes red.

Yes, the failure is displayed below:

 FAIL["test_index_including_pagination", UsersIndexTest,
17.382960474991705]
 test_index_including_pagination#UsersIndexTest (17.38s)
        Expected at least 1 element matching
"a[href="/users/14035331"]", found 0..
        Expected 0 to be >= 1.
        test/integration/users_index_test.rb:15:in `block (2 levels) in
<class:UsersIndexTest>'
        test/integration/users_index_test.rb:14:in `block in
<class:UsersIndexTest>'

=F

Confirm that commenting out only one of the calls to will_paginate
leaves the tests green.

Yes, the test still stay green because the test is only asserting for
one pagination div not two with following code below:

```
 assert_select 'div.pagination'
```
How would you test for the presence of both sets of will_paginate links?
Hint: Use a count from table: Some useful assert_select options.

```
assert_select 'div.pagination', count: 2

``