The first refactoring attempt in the index view.
The paginated users index is now complete, but there’s one improvement I
can’t resist including: Rails has some incredibly slick tools for making
compact views, and in this section we’ll refactor the index page to use
them. Because our code is well-tested, we can refactor with confidence,
assured that we are unlikely to break our site’s functionality.

The first step in our refactoring is to replace the user li from in
commit: The users index with pagination with a render call in this
commit.